requirements.txt
deepeval
anthropic
google-genai
python-dotenv